Remove old gallery screen
